// packages/shared/decorators/src/lib/memoize-async/memoize-async.types.ts
/**
 * Copyright (c) 2026 GDG on Campus Farmingdale State College
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all
 * copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

import type { Cache, KeyResolver, Memoizable } from '../memoize/memoize.types.js';

export type AsyncMemoizable<T, D> = Memoizable<T, Promise<D>>;

export interface AsyncCache<D> {
  set: (key: string, value: D) => Promise<void>;
  get: (key: string) => Promise<D> | Promise<null>;
  delete: (key: string) => Promise<void>;
  has: (key: string) => Promise<boolean>;
}

export interface AsyncMemoizeConfig<T, D> {
  cache?: Cache<D> | AsyncCache<D>;
  keyResolver?: KeyResolver | keyof T;
  expirationTimeMs?: number;
}